RED: Chore: Refactor resource_handler.
